worker/requirements.txt
backports.ssl-match-hostname==3.5.0.1
docker-py==1.9.0
python-etcd==0.4.3
ipaddress==1.0.16
pytz==2013.8
requests==2.11.1
six==1.10.0
websocket-client==0.37.0
